requirements.txt
Pennylane==0.36.0
PyYAML>=6.0.1
scikit-learn==1.3.2
torch==2.5.1
numpy>=1.26.1
tqdm >= 4.65.0